feat: cleaning up old code
