+++ state.py
@@ -1,0 +1,4 @@
+
+class State(object):
+
+    def __init__(self):
